+++ RELEASE_NOTES.md
@@ -1,1 +1,2 @@
  * Added an option to skip the test to see if the host can be resolved
+ * Added an option to resolve an host using DNS over HTTP